Added in the frame of the BattleSystem
Included in this are two scriptable objects, pokemon and moves which can be created within their folders to add more moves and pokemon. Right now there is a basic hud and battle system which displays the player on the bottom left and their enemy on the top right with their stats shown on screen. There is also a dialog box with moves and actions to select (no function right now) which can be highlighted.